Refactor: Extract the Wasm GC array module construction.
